requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
joblib==1.4.2
pandas==2.2.3
Pillow==11.0.0
pydantic==2.10.0
numpy==1.26.4
scikit-learn==1.3.2
python-multipart==0.0.12
# For TFLite inference in production (much lighter than full tensorflow)
tflite-runtime==2.14.0
xgboost==2.1.3
